requirements.txt
# Core framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Email processing
imaplib2==3.6
email-validator==2.1.0
python-email==0.6.0

# Gmail API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Task scheduling and background jobs
celery==5.3.4
redis==5.0.1

# File handling and storage
boto3==1.34.0  # For AWS S3 if using cloud storage
python-multipart==0.0.6

# Data processing
pandas==2.1.3
python-dateutil==2.8.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Logging and monitoring
loguru==0.7.2

# Web interface (optional)
jinja2==3.1.2

# Date/time handling
pytz==2023.3

# Email content parsing
beautifulsoup4==4.12.2
html2text==2020.1.16

# Excel file handling for portfolio data
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1 